+++ src/com/idega/block/creditcard/business/CreditCardBusinessBean.java
@@ -411,7 +411,7 @@ public class CreditCardBusinessBean extends IBOServiceBean implements CreditCard
 		    String str = "";
 		    char[] pass = enc.toCharArray();
 		    for (int i = 0; i < pass.length; i++) {
-		      String hex = Integer.toHexString((int)pass[i]);
+		      String hex = Integer.toHexString(pass[i]);
 		      while (hex.length() < 2) {
 		        String s = "0";
 		        s += hex;
